voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "livekit-agents[bey,google]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "mss>=10.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]
